+++ tests/org.eclipse.n4js.ide.tests/src/org/eclipse/n4js/ide/tests/compiler/AT_GHOLD_212_transpilecrashTest.java
@@ -8,40 +8,36 @@
  * Contributors:
  *   NumberFour AG - Initial API and implementation
  */
-package org.eclipse.n4js.ide.tests;
+package org.eclipse.n4js.ide.tests.compiler;
 
 import static org.eclipse.n4js.cli.N4jscTestOptions.COMPILE;
 import static org.junit.Assert.assertEquals;
 
 import java.io.File;
 import java.io.IOException;
-import java.nio.file.Path;
 
-import org.eclipse.n4js.N4JSGlobals;
 import org.eclipse.n4js.cli.N4jscOptions;
 import org.eclipse.n4js.cli.helper.AbstractCliCompileTest;
 import org.eclipse.n4js.cli.helper.CliCompileResult;
-import org.eclipse.n4js.cli.helper.ProcessResult;
 import org.eclipse.n4js.utils.io.FileDeleter;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 
 /**
- * Test setting final field in the ctor if filled class inherits a spec ctor.
  */
-public class AT_IDEBUG_695_CannotSetFinalFieldInCtorForStaticPolyfillsTest extends AbstractCliCompileTest {
+public class AT_GHOLD_212_transpilecrashTest extends AbstractCliCompileTest {
 
 	File workspace;
-	static String WS_IDEBUG_695 = "IDEBUG-695";
+	static String WSP_212 = "GHOLD-212";
 
 	/**
-	 * Setup workspace.
+	 * Setup workspace with api & api-impl & compile
+	 *
 	 */
 	@Before
 	public void setupWorkspace() throws IOException {
-		workspace = setupWorkspace(WS_IDEBUG_695, false, N4JSGlobals.N4JS_RUNTIME);
-
+		workspace = setupWorkspace(WSP_212, true);
 	}
 
 	/** Delete workspace. */
@@ -50,18 +46,17 @@ public class AT_IDEBUG_695_CannotSetFinalFieldInCtorForStaticPolyfillsTest exten
 		FileDeleter.delete(workspace.toPath(), true);
 	}
 
-	/***/
+	/** The Problem was, that nothing was compiled. */
 	@Test
-	public void compileCheckFinalFieldCanBeSetInInheritedCtor_ExpectCanBeSet() {
-		Path projectDir = workspace.toPath().resolve(WS_IDEBUG_695);
-		Path fileToRun = projectDir.resolve("src-gen/Main.js");
+	public void testCompileOfExtendedIterator_from_RuntimeLibrary() {
 
-		N4jscOptions options = COMPILE(workspace);
+		File proot = new File(workspace, PACKAGES);
+
+		N4jscOptions options = COMPILE(proot);
+
 		CliCompileResult cliResult = n4jsc(options);
-		assertEquals(cliResult.toString(), 2, cliResult.getTranspiledFilesCount());
 
-		ProcessResult nodejsResult = runNodejs(projectDir, fileToRun);
-		assertEquals(nodejsResult.toString(), "A.a == 5: true", nodejsResult.getStdOut());
+		// Make sure, we get here and have exactly one file compiled:
+		assertEquals(cliResult.toString(), 1, cliResult.getTranspiledFilesCount());
 	}
-
 }
