+++ tests/org.eclipse.n4js.ide.tests/src/org/eclipse/n4js/ide/tests/compiler/AT_IDEBUG_654_MissingPolyfillImportsTest.java
@@ -8,40 +8,39 @@
  * Contributors:
  *   NumberFour AG - Initial API and implementation
  */
-package org.eclipse.n4js.ide.tests;
+package org.eclipse.n4js.ide.tests.compiler;
 
 import static org.eclipse.n4js.cli.N4jscTestOptions.COMPILE;
 import static org.junit.Assert.assertEquals;
 
 import java.io.File;
 import java.io.IOException;
+import java.nio.file.Path;
 
 import org.eclipse.n4js.N4JSGlobals;
 import org.eclipse.n4js.cli.N4jscOptions;
 import org.eclipse.n4js.cli.helper.AbstractCliCompileTest;
 import org.eclipse.n4js.cli.helper.CliCompileResult;
-import org.eclipse.n4js.projectModel.names.N4JSProjectName;
+import org.eclipse.n4js.cli.helper.ProcessResult;
 import org.eclipse.n4js.utils.io.FileDeleter;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 
 /**
+ * Test for checking the static polyfill imports in the generated modules.
  */
-public class AT_IDEBUG_532_transpilecrashTest extends AbstractCliCompileTest {
+public class AT_IDEBUG_654_MissingPolyfillImportsTest extends AbstractCliCompileTest {
 
 	File workspace;
-	static String WSP_532 = "IDEBUG-532";
+	static String WS_IDEBUG_654_2 = "IDEBUG-654_2";
 
 	/**
-	 * Setup workspace with api & api-impl & compile
-	 *
+	 * Setup workspace.
 	 */
 	@Before
 	public void setupWorkspace() throws IOException {
-		workspace = setupWorkspace(WSP_532, true,
-				N4JSGlobals.N4JS_RUNTIME, new N4JSProjectName("n4js-runtime-es2015"),
-				new N4JSProjectName("org.eclipse.n4js.mangelhaft"));
+		workspace = setupWorkspace(WS_IDEBUG_654_2, false, N4JSGlobals.N4JS_RUNTIME);
 	}
 
 	/** Delete workspace. */
@@ -50,19 +49,25 @@ public class AT_IDEBUG_532_transpilecrashTest extends AbstractCliCompileTest {
 		FileDeleter.delete(workspace.toPath(), true);
 	}
 
-	/** The Problem was, that nothing was compiled. */
+	/***/
 	@Test
-	public void testCompileOfExtendedIterator_from_RuntimeLibrary() {
+	public void compileCheckPolyfillImports_ExpectExist() {
 
-		File proot = new File(workspace, PACKAGES);
+		Path projectDir = workspace.toPath().resolve(WS_IDEBUG_654_2);
+		Path fileToRun = projectDir.resolve("src-gen/Main.js");
 
-		N4jscOptions options = COMPILE(proot);
-
+		N4jscOptions options = COMPILE(workspace);
 		CliCompileResult cliResult = n4jsc(options);
+		assertEquals(cliResult.toString(), 5, cliResult.getTranspiledFilesCount());
 
-		// Make sure, we get here and have exactly one file compiled:
+		ProcessResult nodejsResult = runNodejs(projectDir, fileToRun);
+		assertEquals(nodejsResult.toString(),
+				"functionFromModuleA\n" +
+						"variableFromModuleB\n" +
+						"variableFromModuleC\n" +
+						"variableFromModuleC",
+				nodejsResult.getStdOut());
 
-		assertEquals(cliResult.toString(), 0, cliResult.getTranspiledFilesCount(proot.toPath().resolve("APIx")));
-		assertEquals(cliResult.toString(), 1, cliResult.getTranspiledFilesCount(proot.toPath().resolve("IMPLx")));
 	}
+
 }
