+++ tests/org.eclipse.n4js.ide.tests/src/org/eclipse/n4js/ide/tests/compiler/AT_IDEBUG_695_CannotSetFinalFieldInCtorForStaticPolyfillsTest.java
@@ -8,36 +8,40 @@
  * Contributors:
  *   NumberFour AG - Initial API and implementation
  */
-package org.eclipse.n4js.ide.tests;
+package org.eclipse.n4js.ide.tests.compiler;
 
 import static org.eclipse.n4js.cli.N4jscTestOptions.COMPILE;
 import static org.junit.Assert.assertEquals;
 
 import java.io.File;
 import java.io.IOException;
+import java.nio.file.Path;
 
+import org.eclipse.n4js.N4JSGlobals;
 import org.eclipse.n4js.cli.N4jscOptions;
 import org.eclipse.n4js.cli.helper.AbstractCliCompileTest;
 import org.eclipse.n4js.cli.helper.CliCompileResult;
+import org.eclipse.n4js.cli.helper.ProcessResult;
 import org.eclipse.n4js.utils.io.FileDeleter;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 
 /**
+ * Test setting final field in the ctor if filled class inherits a spec ctor.
  */
-public class AT_GHOLD_212_transpilecrashTest extends AbstractCliCompileTest {
+public class AT_IDEBUG_695_CannotSetFinalFieldInCtorForStaticPolyfillsTest extends AbstractCliCompileTest {
 
 	File workspace;
-	static String WSP_212 = "GHOLD-212";
+	static String WS_IDEBUG_695 = "IDEBUG-695";
 
 	/**
-	 * Setup workspace with api & api-impl & compile
-	 *
+	 * Setup workspace.
 	 */
 	@Before
 	public void setupWorkspace() throws IOException {
-		workspace = setupWorkspace(WSP_212, true);
+		workspace = setupWorkspace(WS_IDEBUG_695, false, N4JSGlobals.N4JS_RUNTIME);
+
 	}
 
 	/** Delete workspace. */
@@ -46,17 +50,18 @@ public class AT_GHOLD_212_transpilecrashTest extends AbstractCliCompileTest {
 		FileDeleter.delete(workspace.toPath(), true);
 	}
 
-	/** The Problem was, that nothing was compiled. */
+	/***/
 	@Test
-	public void testCompileOfExtendedIterator_from_RuntimeLibrary() {
+	public void compileCheckFinalFieldCanBeSetInInheritedCtor_ExpectCanBeSet() {
+		Path projectDir = workspace.toPath().resolve(WS_IDEBUG_695);
+		Path fileToRun = projectDir.resolve("src-gen/Main.js");
 
-		File proot = new File(workspace, PACKAGES);
-
-		N4jscOptions options = COMPILE(proot);
-
+		N4jscOptions options = COMPILE(workspace);
 		CliCompileResult cliResult = n4jsc(options);
+		assertEquals(cliResult.toString(), 2, cliResult.getTranspiledFilesCount());
 
-		// Make sure, we get here and have exactly one file compiled:
-		assertEquals(cliResult.toString(), 1, cliResult.getTranspiledFilesCount());
+		ProcessResult nodejsResult = runNodejs(projectDir, fileToRun);
+		assertEquals(nodejsResult.toString(), "A.a == 5: true", nodejsResult.getStdOut());
 	}
+
 }
