+++ tests/org.eclipse.n4js.ide.tests/src/org/eclipse/n4js/ide/tests/compiler/InstallCompileRunN4jscExternalWithDefinitionFilesTest.java
@@ -8,7 +8,7 @@
  * Contributors:
  *   NumberFour AG - Initial API and implementation
  */
-package org.eclipse.n4js.ide.tests;
+package org.eclipse.n4js.ide.tests.compiler;
 
 import static org.eclipse.n4js.cli.N4jscTestOptions.COMPILE;
 import static org.junit.Assert.assertEquals;
