+++ tests/org.eclipse.n4js.ide.tests/src/org/eclipse/n4js/ide/tests/compiler/N4jscTypeDefinitionsModuleShadowingTest.java
@@ -8,13 +8,14 @@
  * Contributors:
  *   NumberFour AG - Initial API and implementation
  */
-package org.eclipse.n4js.ide.tests;
+package org.eclipse.n4js.ide.tests.compiler;
 
 import static org.eclipse.n4js.cli.N4jscTestOptions.COMPILE;
 import static org.junit.Assert.assertEquals;
 
 import java.io.File;
 import java.io.IOException;
+import java.util.Collection;
 
 import org.eclipse.n4js.cli.helper.AbstractCliCompileTest;
 import org.eclipse.n4js.cli.helper.CliCompileResult;
@@ -23,21 +24,21 @@ import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 
+import com.google.common.base.Predicates;
+
 /**
- * Tests a project with two project dependencies A,B such that A depends on B.
- *
- * In this test, we have two external projects: nuka-carousel and react. nuka-carousel depends on react via a project
- * dependency, and file P1/src/X.n4jsx depends on both react and nuka-carousel via project dependencies.
+ * Basic tests of type definitions shadowing in the headless case.
  */
-public class N4jscDependentProjectDependenciesTest extends AbstractCliCompileTest {
+public class N4jscTypeDefinitionsModuleShadowingTest extends AbstractCliCompileTest {
+
 	File workspace;
-	File proot;
+	File packages;
 
 	/** Prepare workspace. */
 	@Before
 	public void setupWorkspace() throws IOException {
-		workspace = setupWorkspace("external_project_dependencies", true);
-		proot = new File(workspace, PACKAGES).getAbsoluteFile();
+		workspace = setupWorkspace("type-definitions", Predicates.alwaysFalse(), true);
+		packages = new File(workspace, PACKAGES);
 	}
 
 	/** Delete workspace. */
@@ -46,13 +47,22 @@ public class N4jscDependentProjectDependenciesTest extends AbstractCliCompileTes
 		FileDeleter.delete(workspace.toPath(), true);
 	}
 
-	/** Test failure when compiling without target platform file. */
+	/**
+	 * Builds 'Client', 'Def' and 'Impl' and asserts no issues.
+	 */
 	@Test
-	public void testSuccessfulCompilationWithInterdependentProjects() {
-		yarnInstall(workspace.toPath());
-
+	public void testSimpleTypeDefsShadowing() {
 		CliCompileResult cliResult = n4jsc(COMPILE(workspace));
-		assertEquals(cliResult.toString(), 3, cliResult.getTranspiledFilesCount());
+
+		Collection<String> fileNames = cliResult.getTranspiledFileNames();
+
+		assertEquals(cliResult.toString(), 1, cliResult.getTranspiledFilesCount());
+		assertEquals(cliResult.toString(), "packages/Client/src-gen/Client.js", String.join(", ", fileNames));
+		assertEquals(cliResult.toString(), //
+				"packages/Broken_Client/package.json, "//
+						+ "packages/Broken_Client/src/Client.n4js, "//
+						+ "packages/Client/package.json",
+				String.join(", ", cliResult.getErrFiles()));
 	}
 
 }
